server/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/base.in
#
asgiref==3.4.1
    # via django
bokeh==2.2.3
    # via iotile-analytics-interactive
boto3==1.11.10
    # via
    #   -r requirements/base.in
    #   django-ses
botocore==1.14.17
    # via
    #   boto3
    #   pynamodb
    #   s3transfer
certifi==2020.12.5
    # via
    #   elasticsearch
    #   requests
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
decorator==4.4.2
    # via typedargs
defusedxml==0.6.0
    # via python3-openid
django==3.2.8
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-cors-headers
    #   django-feature-policy
    #   django-filter
    #   django-health-check
    #   django-multiurl
    #   django-recaptcha
    #   django-redshift-backend
    #   django-ses
    #   djangorestframework
    #   drf-yasg
django-allauth==0.42.0
    # via -r requirements/base.in
django-cors-headers==3.4.0
    # via -r requirements/base.in
django-crispy-forms==1.13.0
    # via -r requirements/base.in
django-elasticsearch-dsl==7.1.4
    # via -r requirements/base.in
django-environ==0.4.5
    # via -r requirements/base.in
django-feature-policy==3.4.0
    # via -r requirements/base.in
django-filter==21.1
    # via -r requirements/base.in
django-health-check==3.12.1
    # via -r requirements/base.in
django-multiurl==1.4.0
    # via -r requirements/base.in
# django-pandas==0.6.5
    # via -r requirements/base.in
django-recaptcha==2.0.6
    # via -r requirements/base.in
django-redis==4.4.4
    # via -r requirements/base.in
django-redshift-backend==2.1.0
    # via -r requirements/base.in
django-ses==2.3.0
    # via -r requirements/base.in
djangorestframework==3.11.2
    # via
    #   -r requirements/base.in
    #   drf-yasg
    #   rest-pandas
djangorestframework-api-key==2.0.0
    # via -r requirements/base.in
djangorestframework-jwt==1.11.0
    # via -r requirements/base.in
docutils==0.15.2
    # via botocore
drf-yasg==1.17.1
    # via -r requirements/base.in
elasticsearch==7.10.0
    # via
    #   -r requirements/base.in
    #   elasticsearch-dsl
elasticsearch-dsl==7.3.0
    # via
    #   -r requirements/base.in
    #   django-elasticsearch-dsl
future==0.18.2
    # via
    #   django-ses
    #   iotile-analytics-core
    #   iotile-cloud
    #   structpp
hiredis==1.1.0
    # via -r requirements/base.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
#iotile-analytics-core==0.6.1
    # via iotile-analytics-interactive
iotile-analytics-interactive==0.6.3
    # via -r requirements/base.in
iotile-cloud==0.9.14
    # via
    #   -r requirements/base.in
    #   iotile-analytics-core
itypes==1.2.0
    # via coreapi
jinja2==2.11.2
    # via
    #   bokeh
    #   coreschema
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
markupsafe==1.1.1
    # via jinja2
msgpack==1.0.2
    # via -r requirements/base.in
#numpy==1.19.5
    # via
    #   bokeh
    #   iotile-analytics-core
    #   pandas
    #   scipy
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.8
    # via
    #   bokeh
    #   drf-yasg
#pandas==1.3.4
    # via
    #   -r requirements/base.in
    #   django-pandas
    #   iotile-analytics-core
    #   rest-pandas
pillow==8.1.0
    # via bokeh
psycopg2-binary==2.8.5
    # via -r requirements/base.in
pyjwt==1.7.1
    # via
    #   djangorestframework-jwt
    #   twilio
pynamodb==4.3.3
    # via -r requirements/base.in
pyparsing==2.4.7
    # via packaging
pysocks==1.7.1
    # via twilio
python-dateutil==2.8.1
    # via
    #   bokeh
    #   botocore
    #   elasticsearch-dsl
    #   iotile-cloud
    #   pandas
    #   pynamodb
python3-openid==3.2.0
    # via django-allauth
pytz==2018.7
    # via
    #   -r requirements/base.in
    #   django
    #   django-ses
    #   pandas
    #   twilio
pyyaml==5.4
    # via bokeh
raven==6.10.0
    # via -r requirements/base.in
redis==2.10.6
    # via django-redis
requests==2.25.1
    # via
    #   coreapi
    #   django-allauth
    #   iotile-cloud
    #   requests-aws4auth
    #   requests-oauthlib
    #   twilio
requests-aws4auth==1.0.1
    # via -r requirements/base.in
requests-oauthlib==1.3.0
    # via django-allauth
# rest-pandas==1.1.0
    # via -r requirements/base.in
ruamel.yaml==0.16.12
    # via drf-yasg
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
s3transfer==0.3.4
    # via boto3
scipy==1.5.4
    # via iotile-analytics-core
six==1.15.0
    # via
    #   django-elasticsearch-dsl
    #   django-pandas
    #   drf-yasg
    #   elasticsearch-dsl
    #   pynamodb
    #   python-dateutil
    #   twilio
sqlparse==0.4.1
    # via django
structpp==0.1.0
    # via -r requirements/base.in
tornado==6.1
    # via bokeh
tqdm==4.56.0
    # via iotile-analytics-core
twilio==6.19.2
    # via -r requirements/base.in
typedargs==1.0.1
    # via iotile-analytics-core
typing-extensions==3.7.4.3
    # via bokeh
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.25.11
    # via
    #   botocore
    #   elasticsearch
    #   requests
